Testing binarizing thresholds
Trying out low and high threshold cut-offs for binarizing. Using VIC expt 12 dynamic sph6 as example. Normal threshold is imbinarize(image,0.16). Just tried 0.10, 0.04, 0.22 and 0.28 threshold values. which are+/- multiples of 6 from 0.16
Did not make any manual changes to the binarized images so that comparison can be more accurate, except for smoothing out day 0 image so a boundary can be drawn